Pretty significant advancements. 1. f&f undiscovered constraint that prevented using the same select attrib in different scopes. The effect of which means I can bind a cascading behavior to x selects in a 'global' controler and then I can use the selects in multiple places and the bahavior will hold.
2. checkbox and radio attribs now have a property 'checked' which is a map
of cb/radio indexes to whether that index is checked. You can also give a
cb/radio an id and use that id to index the box on the checked map. The
effect of this, is when combined with the new advanced compiler, you use a
specific checkboxe's checked value as the value for an interpolated
attribute. I use it in a custom element attribute 'hideIf/showIf'.