lots of stuff in flight.
1. I've begun laying out the schema for ads. This reminds me that I should
add an export of the schema to the repo.

2. I've started work on Model.freeze() and Model.thaw(). These will save a
model's attribute state to sessionStorage. The use case is your app
requires a refresh or maybe u just want to be prepared for accidental
ones. Something to help u preserve the UX. I've got a sandboxed test
working but haven't tried it on the real thing.

3. Lastly since I'm on the topic of refreshes, I'm going to resolve the
issue of ending up back @ login page. After investigation I've determined
it's a timeing thing. Export.beginBootstap() is a async callback fired
after Component fetching and I believe it's coming back AFTER the current
URL was fetched.