TAG - new ad form now structured as a series of wrapped partials. The impetus behind this change is I wanted to use the same partial for preview and displaying ads.
To do this I had to move the "forward/back" navigation out. This change
then necistated a way to change botht the action and view of those button.
All these chnages led to the creation of a couple of concepts:

1. extendos (mixins)- dynamically add functionality to an instance of a class. In
this case I can ad "New Ad" specific functions to the Generic Controller
class. The interesting part about this is that the extendo is itself a
static class. This works well when your class needs to interact with
Models or other "global" objects. Puting these objects on the prototype of
the controller is a.) weird and also would require your generic class to
know something about specific Models. This concept allows a generic
Controller clas to ingest and operate on objects it doesn't know exists at
"compile-time".
See this article for some optimization ideas around this concept: https://javascriptweblog.wordpress.com/2011/05/31/a-fresh-look-at-javascript-mixins/

2. Here' I think I've really discovered the powerful nature of using
"global" Models vs. closed scopes (a la AngularJS). Even though the
compiler transforms partials into faux iFrames these "frames" can still
communicate through changing Model states. A concrete example of this is
how the ad form partial controllers change the nav attributes of the
adform model, which changes the UI outside the scope of the "frame"
itself.