Ad editting just about done. Having some issues with changes to images working correctly. There's also a nit-picky issue with the ad category not being re-instated when edit form is loaded. Overall it took me about an hour to set up the re-use of the ad creation forms for ad editting. I can't be certain and I'm obviously bias, but Templar made the process as easy and quick as I think it can be.
The other thing to think about is how much easier making changes to the UI
just got. Change the ad form in one place and it will change everywhere
else. Now there may be a point where the ad creation/editting
functionality begin to diverge. In that case then it won't make sense to
reuse the same partial and controllers but the ability to quickly re-use
large parts of my apps business logic and view in under an hour seems like
it could be a big advantage.