Compiler now annotates nodes that are part of a repeated element with their index. I think this may open the door to targeted re-interpolation of repeats. Also fixed some subtle issues with processing nodes w/ attributes.
I was caching separate nodes even if they were for the same attributes in
different places (ie an different object properties of the same sequential
attribute).

The bug that started this whole process was my destruction of a repeated
cache was blowing up nodes that a.) not part of a repeat (ex. using an
attribs length property outside a repeated ctx) b.) nodes that utilized
components of a repeated attrib (ex. node attrib that is bound to repeat
attrib length).